+++ tap-agent/src/agent/sender_fee_tracker.rs
@@ -34,12 +34,31 @@ pub struct SenderFeeTracker {
     id_to_fee: HashMap<Address, u128>,
     total_fee: u128,
 
+    fees_requesting: u128,
+    ids_requesting: HashSet<Address>,
+
     buffer_window_fee: HashMap<Address, ExpiringSum>,
     buffer_window_duration: Duration,
     // there are some allocations that we don't want it to be
     // heaviest allocation, because they are already marked for finalization,
     // and thus requesting RAVs on their own in their `post_stop` routine.
     blocked_addresses: HashSet<Address>,
+    failed_ravs: HashMap<Address, FailedRavInfo>,
+}
+
+#[derive(Debug, Clone)]
+pub struct FailedRavInfo {
+    failed_ravs_count: u32,
+    failed_rav_backoff_time: Instant,
+}
+
+impl Default for FailedRavInfo {
+    fn default() -> Self {
+        Self {
+            failed_ravs_count: 0,
+            failed_rav_backoff_time: Instant::now(),
+        }
+    }
 }
 
 impl SenderFeeTracker {
@@ -100,9 +119,17 @@ impl SenderFeeTracker {
 
     pub fn get_heaviest_allocation_id(&mut self) -> Option<Address> {
         // just loop over and get the biggest fee
+        let now = Instant::now();
         self.id_to_fee
             .iter()
             .filter(|(addr, _)| !self.blocked_addresses.contains(*addr))
+            .filter(|(addr, _)| !self.ids_requesting.contains(*addr))
+            .filter(|(addr, _)| {
+                self.failed_ravs
+                    .get(*addr)
+                    .map(|failed_rav| now > failed_rav.failed_rav_backoff_time)
+                    .unwrap_or(true)
+            })
             // map to the value minus fees in buffer
             .map(|(addr, fee)| {
                 (
@@ -134,11 +161,11 @@ impl SenderFeeTracker {
     }
 
     pub fn get_total_fee(&self) -> u128 {
-        self.total_fee
+        self.total_fee - self.fees_requesting
     }
 
     pub fn get_total_fee_outside_buffer(&mut self) -> u128 {
-        self.total_fee - self.get_buffer_fee().min(self.total_fee)
+        self.get_total_fee() - self.get_buffer_fee().min(self.total_fee)
     }
 
     pub fn get_buffer_fee(&mut self) -> u128 {
@@ -147,6 +174,31 @@ impl SenderFeeTracker {
             .fold(0u128, |acc, expiring| {
                 acc + expiring.get_sum(&self.buffer_window_duration)
             })
+    }
+
+    pub fn start_rav_request(&mut self, allocation_id: Address) {
+        let current_fee = self.id_to_fee.entry(allocation_id).or_default();
+        self.ids_requesting.insert(allocation_id);
+        self.fees_requesting += *current_fee;
+    }
+
+    /// Should be called before `update`
+    pub fn finish_rav_request(&mut self, allocation_id: Address) {
+        let current_fee = self.id_to_fee.entry(allocation_id).or_default();
+        self.fees_requesting -= *current_fee;
+        self.ids_requesting.remove(&allocation_id);
+    }
+
+    pub fn failed_rav_backoff(&mut self, allocation_id: Address) {
+        // backoff = max(100ms * 2 ^ retries, 60s)
+        let failed_rav = self.failed_ravs.entry(allocation_id).or_default();
+        failed_rav.failed_rav_backoff_time = Instant::now()
+            + (Duration::from_millis(100) * 2u32.pow(failed_rav.failed_ravs_count))
+                .min(Duration::from_secs(60));
+        failed_rav.failed_ravs_count += 1;
+    }
+    pub fn ok_rav_request(&mut self, allocation_id: Address) {
+        self.failed_ravs.remove(&allocation_id);
     }
 }
 
@@ -287,5 +339,67 @@ mod tests {
         assert_eq!(tracker.get_heaviest_allocation_id(), None);
         assert_eq!(tracker.get_total_fee_outside_buffer(), 0);
         assert_eq!(tracker.get_total_fee(), 0);
+    }
+
+    #[test]
+    fn test_filtered_backed_off_allocations() {
+        let allocation_id_0 = address!("abababababababababababababababababababab");
+        let allocation_id_1 = address!("bcbcbcbcbcbcbcbcbcbcbcbcbcbcbcbcbcbcbcbc");
+        const BACK_SLEEP_DURATION: Duration = Duration::from_millis(201);
+
+        let mut tracker = SenderFeeTracker::default();
+        assert_eq!(tracker.get_heaviest_allocation_id(), None);
+        assert_eq!(tracker.get_total_fee(), 0);
+
+        tracker.update(allocation_id_0, 10);
+        assert_eq!(tracker.get_heaviest_allocation_id(), Some(allocation_id_0));
+        assert_eq!(tracker.get_total_fee(), 10);
+
+        tracker.update(allocation_id_1, 20);
+        assert_eq!(tracker.get_heaviest_allocation_id(), Some(allocation_id_1));
+        assert_eq!(tracker.get_total_fee(), 30);
+
+        // Simulate failed rav and backoff
+        tracker.failed_rav_backoff(allocation_id_1);
+
+        // Heaviest should be the first since its not blocked nor failed
+        assert_eq!(tracker.get_heaviest_allocation_id(), Some(allocation_id_0));
+        assert_eq!(tracker.get_total_fee(), 30);
+
+        sleep(BACK_SLEEP_DURATION);
+
+        assert_eq!(tracker.get_heaviest_allocation_id(), Some(allocation_id_1));
+        assert_eq!(tracker.get_total_fee(), 30);
+    }
+
+    #[test]
+    fn test_ongoing_rav_requests() {
+        let allocation_id_0 = address!("abababababababababababababababababababab");
+        let allocation_id_1 = address!("bcbcbcbcbcbcbcbcbcbcbcbcbcbcbcbcbcbcbcbc");
+        let allocation_id_2 = address!("cdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcd");
+
+        let mut tracker = SenderFeeTracker::default();
+        assert_eq!(tracker.get_heaviest_allocation_id(), None);
+        assert_eq!(tracker.get_total_fee_outside_buffer(), 0);
+        assert_eq!(tracker.get_total_fee(), 0);
+
+        tracker.add(allocation_id_0, 10);
+        tracker.add(allocation_id_1, 20);
+        tracker.add(allocation_id_2, 30);
+        assert_eq!(tracker.get_heaviest_allocation_id(), Some(allocation_id_2));
+        assert_eq!(tracker.get_total_fee(), 60);
+        assert_eq!(tracker.get_total_fee_outside_buffer(), 60);
+
+        tracker.start_rav_request(allocation_id_2);
+
+        assert_eq!(tracker.get_heaviest_allocation_id(), Some(allocation_id_1));
+        assert_eq!(tracker.get_total_fee(), 30);
+        assert_eq!(tracker.get_total_fee_outside_buffer(), 30);
+
+        tracker.finish_rav_request(allocation_id_2);
+
+        assert_eq!(tracker.get_heaviest_allocation_id(), Some(allocation_id_2));
+        assert_eq!(tracker.get_total_fee(), 60);
+        assert_eq!(tracker.get_total_fee_outside_buffer(), 60);
     }
 }